third_party/WebKit/LayoutTests/intersection-observer/shadow-content-expected.txt

Tests that observations can be made across shadow DOM boundaries.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS entries.length is 0
PASS entries.length is 1
PASS entries[0].boundingClientRect.left is 8
PASS entries[0].boundingClientRect.right is 108
PASS entries[0].boundingClientRect.top is 8
PASS entries[0].boundingClientRect.bottom is 108
PASS entries[0].intersectionRect.left is 8
PASS entries[0].intersectionRect.right is 108
PASS entries[0].intersectionRect.top is 8
PASS entries[0].intersectionRect.bottom is 108
PASS entries[0].rootBounds.left is 0
PASS entries[0].rootBounds.right is 800
PASS entries[0].rootBounds.top is 0
PASS entries[0].rootBounds.bottom is 600
PASS entries[0].target is [object HTMLDivElement]
PASS successfullyParsed is true

TEST COMPLETE
